# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.FriendInfoResp:
    properties:
      friend_remark:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      uid:
        type: string
    type: object
...
  model.LoginByPhoneReq:
    properties:
      password:
        maxLength: 72
        minLength: 6
        type: string
      phone_number:
        type: string
    required:
    - password
    - phone_number
    type: object
  model.LoginByUidReq:
    properties:
      password:
        maxLength: 72
        minLength: 6
        type: string
      uid:
        maxLength: 20
        minLength: 1
        type: string
    required:
    - password
    - uid
    type: object
  model.LoginResp:
    properties:
      token_class:
        $ref: '#/definitions/model.TokenResp'
      user_info:
        $ref: '#/definitions/model.UserInfoResp'
    type: object
  model.NameReq:
    properties:
      name:
        maxLength: 64
        minLength: 1
        type: string
    required:
    - name
    type: object
  model.RegisterReq:
    properties:
      name:
        maxLength: 64
        minLength: 1
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
      phone_number:
        type: string
    required:
    - name
    - password
    - phone_number
    type: object
...
  model.RevisePasswordReq:
    properties:
      new_password:
        maxLength: 72
        minLength: 6
        type: string
      prev_password:
        maxLength: 72
        minLength: 6
        type: string
    required:
    - new_password
    - prev_password
    type: object
...
  model.StrangerInfoResp:
    properties:
      id:
        example: "0"
        type: string
      name:
        type: string
    type: object
  model.TokenResp:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
    type: object
  model.UidReq:
    properties:
      uid:
        maxLength: 20
        minLength: 1
        type: string
    required:
    - uid
    type: object
  model.UserInfoResp:
    properties:
      name:
        type: string
      uid:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Inquire backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Inquire API
  version: "1.0"
paths:
...
  /auth/info/friends/id/{id}:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取好友详细信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 好友用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.FriendInfoResp'
              type: object
        "400":
          description: ID错误或找不到好友
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 根据ID查看好友信息
      tags:
      - user
  /auth/info/friends/uid/{uid}:
    get:
      consumes:
      - application/json
      description: 根据用户Uid获取好友详细信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 好友Uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.FriendInfoResp'
              type: object
        "400":
          description: Uid为空或找不到好友
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 根据Uid查看好友信息
      tags:
      - user
  /auth/info/strangers/id/{id}:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取陌生人信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 陌生人用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.StrangerInfoResp'
              type: object
        "400":
          description: ID错误或找不到此人
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 根据ID查看陌生人信息
      tags:
      - user
  /auth/info/strangers/uid/{uid}:
    get:
      consumes:
      - application/json
      description: 根据用户Uid获取陌生人信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 陌生人Uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.StrangerInfoResp'
              type: object
        "400":
          description: Uid为空或找不到此人
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 根据Uid查看陌生人信息
      tags:
      - user
  /auth/me/name:
    post:
      consumes:
      - application/json
      description: 修改当前用户的用户名
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 修改用户名请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.NameReq'
      produces:
      - application/json
      responses:
        "201":
          description: 修改成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改用户名
      tags:
      - user
  /auth/me/password:
    post:
      consumes:
      - application/json
      description: 修改当前用户的密码
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 修改密码请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.RevisePasswordReq'
      produces:
      - application/json
      responses:
        "201":
          description: 修改成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改密码
      tags:
      - user
  /auth/me/uid:
    post:
      consumes:
      - application/json
      description: 修改当前用户的微信号
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 修改微信号请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.UidReq'
      produces:
      - application/json
      responses:
        "201":
          description: 修改成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误或微信号重复
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改微信号
      tags:
      - user
...
  /auth/refresh_token:
    post:
      consumes:
      - application/json
      description: 使用 Refresh Token 获取新的 Access Token
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TokenResp'
              type: object
        "500":
          description: Token问题
          schema:
            $ref: '#/definitions/response.Response'
      summary: 刷新Token
      tags:
      - auth
  /login/phone_number:
    post:
      consumes:
      - application/json
      description: 使用手机号和密码登录
      parameters:
      - description: 手机号登录请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.LoginByPhoneReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LoginResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 手机号登录
      tags:
      - user
  /login/uid:
    post:
      consumes:
      - application/json
      description: 使用微信号和密码登录
      parameters:
      - description: 微信号登录请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.LoginByUidReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LoginResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 微信号登录
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 注册请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误或手机号重复
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"